--- a/.rspec
+++ b/.rspec
@@ -1,1 +1,5 @@
 --require spec_helper
+--require rails_helper
+--format progress
+--color
+--order random