--- a/.rspec-integration
+++ b/.rspec-integration
@@ -1,0 +1,7 @@
+--require spec_helper
+--require rails_helper
+--format documentation
+--color
+--order random
+--tag integration
+--exclude-pattern "spec/{unit,performance,system}/**/*_spec.rb"