--- a/Procfile.dev
+++ b/Procfile.dev
@@ -1,4 +1,4 @@
-web: bin/rails server -p 3001
+web: bin/rails server
 css: bin/rails tailwindcss:watch
 worker: bin/rails solid_queue:start
 sidekiq: bundle exec sidekiq