--- a/bin/test-performance
+++ b/bin/test-performance
@@ -1,0 +1,39 @@
+#!/usr/bin/env ruby
+# Performance tests for benchmarking and regression detection
+# Only runs when explicitly requested
+
+require 'fileutils'
+
+# Ensure we're in the Rails root
+Dir.chdir(File.expand_path('..', __dir__))
+
+# Set test environment
+ENV['RAILS_ENV'] = 'test'
+
+# Build RSpec command for performance tests
+cmd_parts = [
+  'bundle exec rspec',
+  '--tag performance',
+  '--format documentation',
+  '--order defined'  # Maintain consistent order for performance comparison
+]
+
+cmd = cmd_parts.join(' ')
+
+puts "⚡ Running performance tests (may take several minutes)..."
+puts "Command: #{cmd}"
+puts
+
+# Execute with proper exit code
+system(cmd)
+exit_code = $?.exitstatus
+
+if exit_code == 0
+  puts
+  puts "✅ All performance tests passed!"
+else
+  puts
+  puts "❌ Some performance tests failed (exit code: #{exit_code})"
+end
+
+exit exit_code