--- a/spec/models/admin_user_spec.rb
+++ b/spec/models/admin_user_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe AdminUser, type: :model do
+RSpec.describe AdminUser, type: :model, unit: true do
   pending "add some examples to (or delete) #{__FILE__}"
 end