… .github/workflows/test-suite.yml
… Gemfile.lock
… ORCHESTRATOR_IMPROVEMENTS_SUMMARY.md
… app/services/categorization/orchestrator.rb
… app/services/categorization/orchestrator_factory.rb
… app/services/categorization/pattern_learner.rb
… bin/test-verify
… config/database.yml
… config/routes.rb
… demo_orchestrator_improvements.rb
… docs/categorization_improvement/PHASE_1_GAP_ANALYSIS.md
… docs/categorization_improvement/option1_tasks/task-1.7.2-enhanced.md
… docs/categorization_improvement/option1_tasks/task-1.7.3-enhanced.md
… docs/categorization_improvement/option1_tasks/task-1.7.4-enhanced.md
… docs/implementation/IMPLEMENTATION_GUIDE.md
… docs/implementation/tickets/epic-1-foundation/ticket-1.1-setup-gems-migrations.md
… docs/implementation/tickets/epic-1-foundation/ticket-1.2-account-user-models.md
… docs/implementation/tickets/epic-1-foundation/ticket-1.3-authentication-setup.md
… docs/implementation/tickets/epic-1-foundation/ticket-1.4-account-invitation-system.md
… docs/implementation/tickets/epic-2-multi-tenancy/ticket-2.1-tenant-scoping-models.md
… docs/implementation/tickets/epic-2-multi-tenancy/ticket-2.2-update-controllers-tenant-context.md
… docs/implementation/tickets/epic-2-multi-tenancy/ticket-2.3-update-services-layer.md
… docs/implementation/tickets/epic-3-privacy-features/ticket-3.1-expense-visibility-system.md
… docs/implementation/tickets/epic-3-privacy-features/ticket-3.2-role-based-permissions.md
… docs/implementation/tickets/epic-4-polish-migration/ticket-4.1-data-migration-script.md
… docs/implementation/tickets/epic-4-polish-migration/ticket-4.2-ui-polish-account-management.md
… docs/implementation/tickets/epic-4-polish-migration/ticket-4.3-comprehensive-testing-suite.md
… docs/implementation/tickets/implementation-summary.md
… docs/multi-tenancy-implementation.md
… docs/requirements/multi-user-accounts-rbac.md